fix: move store to make apparent it belongs to the hook & add comments
