+++ src/Molecules/Tooltip/Tooltip.tsx
@@ -4,6 +4,18 @@ import { TooltipPopup } from './TooltipPopup';
 import { wrapEvent, mergeRefs } from '../../common/utils';
 import { useTooltip } from './hooks';
 
+/** 
+  There are a few features that are important to understand.
+  
+  1. Tooltips don't show up until the user has rested on one, we don't
+     want tooltips popupping up as you move your mouse around the page.
+  
+  2. Once any tooltip becomes visible, other tooltips nearby should skip
+     resting and display immediately.
+  
+  3. Tooltips stick around for a little bit after blur/mouseleave. 
+*/
+
 export const Tooltip: TooltipComponent = forwardRef(
   (
     { children, label, ariaLabel, position = 'bottom', mode = 'hover', inModal, maxWidth = 50 },
